test: Hande return value type properly
t/xs.t test failed on big-endian. The return_value after ffi_call()
was always 0. Big endian systems have the least siginicant byte at the
end of a word and interpreting a signed char memory as an integer
must take that offset into account.

This was not the case of the t/xs.t where doublechaar() returned
unsigned char but return_value was of int type and thus libffi stored
the value at a wrong offset.

Moreover the libffi was not properly used because a storage size of a libffi
return value must always be at least sizeof(ffi_arg) large and that's
not true for an unsigned char variable.

This patch fixes it.

Signed-off-by: Petr Písař <ppisar@redhat.com>